# Optimize provider switching performance with batched storage operations