--- a/.changeset/purple-rings-hide.md
+++ b/.changeset/purple-rings-hide.md
@@ -1,0 +1,5 @@
+---
+"claude-dev": patch
+---
+
+Optimizing provider switching performance with batched storage operations