### Description

This PR dramatically improves provider switching performance by optimizing storage operations, reducing switching time from ~550ms to ~30ms (18x improvement).

**Problem:** Provider switching was extremely slow due to inefficient storage operations. Performance analysis revealed that switching providers triggered 70+ individual VSCode storage API calls:
- 47 sequential `updateGlobalState()` calls taking ~527ms


**Solution:** Implemented batched storage operations using `Promise.all()` to reduce API overhead:
- Consolidated 47 global state updates into 1 batched operation
- Consolidated 24 secret updates into 1 batched operation  
- Execute both batches in parallel for maximum performance

**Impact:** Provider switching now feels instant to users, improving the overall UX significantly.

**Performance Logs - Before Optimization:**
```
[PERF] Global state updates took: 307.86ms
[PERF] Ephemeral model config updates took: 219.65ms  
[PERF] Total updateApiConfiguration took: 535.62ms
```

**Performance Logs - After Optimization:**
```
[PERF] Batched global state update took: 17.52ms
[PERF] Batched secrets update took: 9.06ms
[PERF] OPTIMIZED updateApiConfiguration took: 18.17ms (was ~535ms)
[PERF] Total updateApiConfigurationProto took: 29.98ms
```

### Test Procedure

**Performance Testing:**
1. Added comprehensive timing measurements throughout the pipeline
2. Tested provider switching before/after optimization using browser console logs
3. Verified 18x performance improvement (550ms → 30ms) across multiple provider switches
4. Confirmed all existing functionality remains intact

**Compatibility Testing:**
- Verified backward compatibility - all existing APIs unchanged
- Tested with multiple provider types (Anthropic, OpenRouter, OpenAI, etc.)
- Confirmed no breaking changes to storage interface
- Validated that all provider-specific settings are preserved correctly

**Code Quality:**
- Added detailed performance logging for future debugging
- Maintained existing error handling patterns
- Ensured type safety throughout the changes

### Type of Change

- [x] ♻️ Refactor Changes
- [x] 🐛 Bug fix (non-breaking change which fixes an issue)

### Pre-flight Checklist

- [x] Changes are limited to a single feature, bugfix or chore (split larger changes into separate PRs)
- [x] Tests are passing (`npm test`) and code is formatted and linted (`npm run format && npm run lint`)
- [x] I have created a changeset using `npm run changeset` (required for user-facing changes)
- [x] I have reviewed [contributor guidelines](https://github.com/cline/cline/blob/main/CONTRIBUTING.md)

### Screenshots



### Additional Notes

**Files Modified:**
- `src/core/storage/state.ts` - Added batching utilities and refactored `updateApiConfiguration()`
- `src/core/controller/models/updateApiConfigurationProto.ts` - Added performance monitoring
- `webview-ui/src/components/settings/utils/useApiConfigurationHandlers.ts` - Added frontend timing
- `PERFORMANCE_OPTIMIZATION_SUMMARY.md` - Technical documentation

**Technical Details:**
The optimization leverages `Promise.all()` to batch VSCode storage operations, reducing the per-call overhead that was causing the performance bottleneck. This maintains full backward compatibility while providing dramatic performance improvements.

<!-- ELLIPSIS_HIDDEN -->


----

> [!IMPORTANT]
> Optimizes provider switching by batching storage operations in `state.ts`, improving performance from 550ms to 30ms.
> 
>   - **Performance Optimization**:
>     - Introduced `updateGlobalStateBatch()` and `updateSecretsBatch()` in `state.ts` to batch storage operations using `Promise.all()`.
>     - Refactored `updateApiConfiguration()` in `state.ts` to use batched operations, reducing 47 global state updates to 2 and 23 secret updates to 1.
>     - Achieved 18x performance improvement in provider switching (550ms to 30ms).
>   - **Testing and Compatibility**:
>     - Verified performance improvements and backward compatibility across multiple provider types.
>     - Ensured no breaking changes to existing APIs and preserved provider-specific settings.
>   - **Miscellaneous**:
>     - Added performance logging in `updateApiConfigurationProto.ts`.
>     - Updated frontend timing in `useApiConfigurationHandlers.ts`.
> 
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=cline%2Fcline&utm_source=github&utm_medium=referral)<sup> for 4dd9f97254b249eccfe824baeb4c3d89137123a6. You can [customize](https://app.ellipsis.dev/cline/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>


<!-- ELLIPSIS_HIDDEN -->