@arafatkatze 2025-07-06
> I would ask why we need to be renaming the function and variables instead of just adding Promise.all()

There's not necessarily a renaming but just bucketing and classification of different kinds of variables makes it much easier to console log and also its easier to read. As long as we keep the `promise.all` I don't really care eitherway. If you feel there's a cleaner solution feel free to just push on this branch. 

> you can see that getAllExtensionState uses Promise.all() implicitly.

Fair