@celestial-vault 2025-07-07
Yeah I think after the logging is done, just adding Promise.all() should do the trick, no?